requirements.txt
apache-airflow
streamlit
altair==4.0
boto3
pandas
altair
matplotlib
seaborn